use existing uuid if it exists
the previous query only added a uuid to the public graph if the subject did
not have a uuid in any graph. Since the sync service generates uuids this
caused some subjects to have no uuids in the public graph.

the query was modified to generate a uuid if there isn't one defined yet and
to copy the existing one if present.